fix: change 'next' text color
The original cyan color looked nice enough, but was too different from
the standard grey for a normal task.  I still wanted next actions to
stand out, but the cyan was also looking bad on some of the background
colors.

So hopefully making it white will allow it to stand out, but not look
too different in a standard task, and show up better on the styles that
have a background color